Config can now be generated by js
